+++ extensions/anybox/src/components/FolderItem.tsx
@@ -1,5 +1,6 @@
-import { ActionPanel, Action, List, Icon } from "@raycast/api";
+import { ActionPanel, Action, List, Grid, Icon, getPreferenceValues } from "@raycast/api";
 import { FolderProp } from "../utilities/fetch";
+import { Preferences } from "../utilities/searchRequest";
 import SearchFolder from "../searchFolder";
 
 interface Props {
@@ -8,29 +9,58 @@ interface Props {
 
 export default function FolderItem(props: Props) {
   const folder = props.item;
-  return (
-    <List.Item
-      title={folder.name}
-      icon={{
-        source: `http://127.0.0.1:6391/sf-symbols/${folder.icon}`,
-        fallback: Icon.Folder,
-        tintColor: folder.color,
-      }}
-      accessories={[{ text: "Folder" }]}
-      id={folder.id}
-      actions={
-        <ActionPanel title={`${folder.name}`}>
-          <Action.Push
-            title="Show Links in Tag"
-            icon={{
-              source: `http://127.0.0.1:6391/sf-symbols/${folder.icon}`,
-              fallback: Icon.Folder,
-              tintColor: folder.color,
-            }}
-            target={<SearchFolder folder={folder} />}
-          ></Action.Push>
-        </ActionPanel>
-      }
-    />
-  );
+  const preferences: Preferences = getPreferenceValues();
+  if (preferences.asIcons) {
+    return (
+      <Grid.Item
+        title={folder.name}
+        content={{
+          source: `http://127.0.0.1:6391/sf-symbols/${folder.icon}`,
+          fallback: Icon.Folder,
+          tintColor: folder.color,
+        }}
+        subtitle="Folder"
+        id={folder.id}
+        actions={
+          <ActionPanel title={`${folder.name}`}>
+            <Action.Push
+              title="Show Links in Tag"
+              icon={{
+                source: `http://127.0.0.1:6391/sf-symbols/${folder.icon}`,
+                fallback: Icon.Folder,
+                tintColor: folder.color,
+              }}
+              target={<SearchFolder folder={folder} />}
+            ></Action.Push>
+          </ActionPanel>
+        }
+      />
+    );
+  } else {
+    return (
+      <List.Item
+        title={folder.name}
+        icon={{
+          source: `http://127.0.0.1:6391/sf-symbols/${folder.icon}`,
+          fallback: Icon.Folder,
+          tintColor: folder.color,
+        }}
+        accessories={[{ text: "Folder" }]}
+        id={folder.id}
+        actions={
+          <ActionPanel title={`${folder.name}`}>
+            <Action.Push
+              title="Show Links in Tag"
+              icon={{
+                source: `http://127.0.0.1:6391/sf-symbols/${folder.icon}`,
+                fallback: Icon.Folder,
+                tintColor: folder.color,
+              }}
+              target={<SearchFolder folder={folder} />}
+            ></Action.Push>
+          </ActionPanel>
+        }
+      />
+    );
+  }
 }