+++ extensions/deepcast/src/korean.ts
@@ -1,7 +1,6 @@
+import { LaunchProps } from "@raycast/api";
 import { translate } from "./utils";
 
-const Korean = async () => {
-  await translate("KO");
-  return null;
-};
-export default Korean;
+export default async function Korean(props: LaunchProps<{ arguments?: Arguments.Korean }>) {
+  await translate("KO", props.arguments?.text ?? props.fallbackText);
+}