+++ extensions/password-generator/src/helpers/helpers.ts
@@ -1,3 +1,5 @@
+import crypto from "node:crypto";
+
 export function generatePassword(len: number, useNumbers: boolean, useChars: boolean): string {
   let charset = "abcdefghijkmnopqrstuvwxyzABCDEFGHJKLMNPQRSTUVWXYZ";
   if (useNumbers) {
@@ -7,8 +9,8 @@ export function generatePassword(len: number, useNumbers: boolean, useChars: boo
     charset += "!@#$*^&%";
   }
   let retVal = "";
-  for (let i = 0, n = charset.length; i < len; ++i) {
-    retVal += charset.charAt(Math.floor(Math.random() * n));
+  for (let i = 0; i < len; ++i) {
+    retVal += charset.charAt(crypto.randomInt(charset.length));
   }
   return retVal;
 }