+++ extensions/reclaim-ai/CHANGELOG.md
@@ -1,5 +1,16 @@
 # reclaim Changelog
 
+## [Update] - 2023-10-31
+
+- Refactor of Event Fetching Process
+- Bug Fix in the Menu Bar Component
+
+## [Update] - 2023-10-24
+
+- Revert filter logic
+- New prioritization features
+- Bug fixes
+
 ## [Update] - 2023-10-11
 
 - Show All Connected Calendar Events: Updated the system to display events from all calendars connected via Reclaim.